# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aiverse API",
    "description": "API documentation for the Aiverse autonomous e-commerce platform",
    "version": "1.0.0",
    "contact": {
      "name": "Aiverse Support",
      "url": "https://alverse.app/contact"
    }
  },
  "servers": [
    {
      "url": "https://alverse.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
    }
  },
  "paths": {
...
    "/api/world/halls/{id}": {
      "get": {
        "summary": "Get hall details",
        "description": "Returns detailed information about a specific hall",
        "tags": ["World"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Hall ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Hall data retrieved successfully"
          },
          "404": {
            "description": "Hall not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/world/streets/{slug}": {
      "get": {
        "summary": "Get street details",
        "description": "Returns detailed information about a specific street",
        "tags": ["World"],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Street slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Street data retrieved successfully"
          },
          "404": {
            "description": "Street not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/world/evolution": {
      "post": {
        "summary": "Trigger world evolution",
        "description": "Manually trigger the autonomous world evolution process",
        "tags": ["World", "Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Evolution triggered successfully"
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/auto-listing/extract": {
      "post": {
        "summary": "Extract product data",
        "description": "Extracts product information from external URLs",
        "tags": ["Auto Listing"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_url": { "type": "string" },
                  "supplier_id": { "type": "string" }
                },
                "required": ["product_url", "supplier_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product data extracted successfully"
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
...
}